feat(runloop) rebuild plugins_iterator and router via a timer
This change creates a timer which checks and rebuilds the plugins iterator and router, once per second, so the rebuilding is partially uncoupled from receiving a request.

Extracted from #4488, with some minor differences:

* `rebuild_sync` was renamed to `safe_rebuild`
* Some functions have changed so that they return nil, err instead of
logging the error and just returning false
* The `rebuild` utility function now returns nil, err as well, it is up
to the caller to decide how to deal with errors.
* The `wait` parameter has been renamed to `timeout` and is mandatory
now.
* The `acquire_semaphore` function now returns `ok, err`, and the err
can be just `timeout`
* Timer-based rebuilds explicitly set the `timeout` to 0 and simply
ignore the case where `acquire_semaphore` returns `timeout`
* `get_plugins_iterator` gets the iterator, while
`get_updated_plugins_iterator` attempts to rebuild it first. In case of
failure, it logs it and returns the old plugins iterator.

Co-Authored-By: Aapo Talvensaari <bungle@konghq.com>