feat(runloop) add db.connect/close/keepalive to retries to rebuilds
This commit attempts to solve several problems which happen only under heavy load. They are detailed in #4194. This is extracted from #4488

The idea is that when the callback (update_router or
update_plugins_iterator) fails, instead of just returning the failure,
we retry again, but by wrapping the callback with `db:connect` and
`db:keepalive`.

This seems to have two benefits:
* It solves some sporadic NPEs which seem to occur on timers during heavy load
* It speeds up rebuilds in Cassandra