feat(*) support custom logging serializers
This commit introduces a new DAO entity, 'log_serializers', to
support logging plugins writing output in customizable, arbitrary
format. The log_serializers objects are comprised of identifiers
(ID, name) and a chunk field that shall contain a base64-encoded
Lua string. Chunks are expected to expose a function 'serialize'
that returns data to be written by the logging plugins. Plugins
themselves are given a new config field 'serializer', denoting the
name of the serializer that should be used on the plugin. The
legacy serializer module at kong.plugins.log-serializers.basic
is kept in the project for backwards compatability.